# frostdb: add leveled compaction