--- a/dynparquet/row.go
+++ b/dynparquet/row.go
@@ -42,6 +42,10 @@ type DynamicRow struct {
 }
 
 func (s *Schema) RowLessThan(a, b *DynamicRow) bool {
+	return s.Cmp(a, b) < 0
+}
+
+func (s *Schema) Cmp(a, b *DynamicRow) int {
 	dynamicColumns := mergeDynamicColumnSets([]map[string][]string{a.DynamicColumns, b.DynamicColumns})
 	cols := s.parquetSortingColumns(dynamicColumns)
 	for _, col := range cols {
@@ -64,16 +68,16 @@ func (s *Schema) RowLessThan(a, b *DynamicRow) bool {
 		av, bv := extractValues(a, b, aIndex, bIndex)
 		cmp := compare(col, node, av, bv)
 		if cmp < 0 {
-			return true
+			return cmp
 		}
 		if cmp > 0 {
-			return false
+			return cmp
 		}
 		// neither of those case are true so a and b are equal for this column
 		// and we need to continue with the next column.
 	}
 
-	return false
+	return 0
 }
 
 func compare(col parquet.SortingColumn, node parquet.Node, av, bv []parquet.Value) int {