This PR implements simple leveled compaction of parts into level 1 parts as described in #223, which are guaranteed to never be overlapping. This will hopefully improve our memory usage during compactions and avoid the spikes we've seen in our clusters. Please refer to the individual commits for details. I've tried to split it up into 1) The addition of leveled compaction, 2) Handling out-of-order inserts which should only happen very rarely but it is an edge case we need to handle.