@asubiotto 2022-10-20
By "much worse" I was referring to benchmark performance. However I think this compaction strategy already makes things better. Based on my investigation, we were never actually running a compaction on main during these benchmarks. This is because we would never unset `pruned` on a granule in cases where only parts with uncompleted txns were found (this is the case of compacting during WAL replays). Since `pruned` was already set, the granule would never be compacted.

It turns out the inserts are actually very well batched. Here is a comparison of main with and without our current compaction strategy (simply adding a `t.abort(granule)` in the case of `len(bufs) == 0` in `compactGranule`):
```
name           old time/op    new time/op     delta
Query/Types-8    3.77ms ± 1%   641.41ms ±93%  +16929.55%  (p=0.000 n=10+10)
Query/Merge-8    12.4ms ± 1%     27.9ms ± 2%    +124.03%  (p=0.000 n=9+9)
Query/Range-8    4.04ms ± 0%    14.10ms ± 1%    +249.33%  (p=0.000 n=8+10)

name           old alloc/op   new alloc/op    delta
Query/Types-8    5.69MB ± 0%  5353.70MB ±91%  +93949.13%  (p=0.000 n=10+10)
Query/Merge-8    74.9MB ± 0%    162.8MB ± 0%    +117.36%  (p=0.000 n=9+10)
Query/Range-8    16.3MB ± 0%     58.5MB ± 0%    +258.73%  (p=0.000 n=10+10)

name           old allocs/op  new allocs/op   delta
Query/Types-8     93.4k ± 0%   2495.5k ±359%        ~     (p=0.720 n=10+9)
Query/Merge-8      301k ± 0%       204k ± 0%     -32.32%  (p=0.000 n=9+10)
Query/Range-8     57.7k ± 0%       7.8k ± 0%     -86.49%  (p=0.000 n=10+10)
```

Now compare the leveled compaction strategy with the our current compaction strategy:
```
name           old time/op    new time/op    delta
Query/Types-8     641ms ±93%      42ms ± 6%  -93.46%  (p=0.000 n=10+10)
Query/Merge-8    27.9ms ± 2%    22.3ms ± 1%  -19.81%  (p=0.000 n=9+10)
Query/Range-8    14.1ms ± 1%    10.3ms ± 0%  -27.25%  (p=0.000 n=10+10)

name           old alloc/op   new alloc/op   delta
Query/Types-8    5.35GB ±91%    0.39GB ± 0%  -92.63%  (p=0.000 n=10+10)
Query/Merge-8     163MB ± 0%     119MB ± 0%  -26.90%  (p=0.000 n=10+10)
Query/Range-8    58.5MB ± 0%    36.7MB ± 0%  -37.27%  (p=0.000 n=10+10)

name           old allocs/op  new allocs/op  delta
Query/Types-8    2.50M ±359%     0.02M ± 0%  -99.33%  (p=0.004 n=9+9)
Query/Merge-8      204k ± 0%      203k ± 0%   -0.13%  (p=0.000 n=10+10)
Query/Range-8     7.80k ± 0%     7.54k ± 0%   -3.32%  (p=0.000 n=10+10)
```

The leveled compaction strategy still performs worse than no compactions but is definitely better than the current compaction strategy. For reference here is no compactions vs leveled compaction:
```
name           old time/op    new time/op    delta
Query/Types-8    3.77ms ± 1%   41.94ms ± 6%  +1013.38%  (p=0.000 n=10+10)
Query/Merge-8    12.4ms ± 1%    22.3ms ± 1%    +79.65%  (p=0.000 n=9+10)
Query/Range-8    4.04ms ± 0%   10.26ms ± 0%   +154.13%  (p=0.000 n=8+10)

name           old alloc/op   new alloc/op   delta
Query/Types-8    5.69MB ± 0%  394.45MB ± 0%  +6829.39%  (p=0.000 n=10+10)
Query/Merge-8    74.9MB ± 0%   119.0MB ± 0%    +58.89%  (p=0.000 n=9+10)
Query/Range-8    16.3MB ± 0%    36.7MB ± 0%   +125.03%  (p=0.000 n=10+10)

name           old allocs/op  new allocs/op  delta
Query/Types-8     93.4k ± 0%     16.7k ± 0%    -82.18%  (p=0.000 n=10+9)
Query/Merge-8      301k ± 0%      203k ± 0%    -32.41%  (p=0.000 n=9+10)
Query/Range-8     57.7k ± 0%      7.5k ± 0%    -86.94%  (p=0.000 n=10+10)
```

My explanation for this is that `QueryTypes` worsened a lot because distinct performance improvements cannot be taken advantage of after reshaping the data during compaction (still to be confirmed). Merge and Range I am not quite sure of, but it's probably due to a similar explanation.

I think that I will add some improvements to this PR that I thought about when replaying WALs (writing l1s iteratively from a single merged row group of all l0s) and advocate to merge this given its improvement over our current compaction strategy.

It's clear that we have to do more work when it comes to "shaping" parts so that the scan level can reach the non-compacted latencies, but that is left for future work. The important part is that this compaction is an improvement over our current compaction strategy.