@asubiotto 2022-10-20
This is now ready for a look! Add a commit with the new "cursor" strategy which simplifies things a bit. The performance hit from no compactions has also been mitigated:
```
name           old time/op    new time/op    delta
Query/Types-8    3.77ms ± 1%   11.07ms ± 3%   +193.88%  (p=0.000 n=10+10)
Query/Merge-8    12.4ms ± 1%    20.4ms ± 0%    +63.69%  (p=0.000 n=9+7)
Query/Range-8    4.04ms ± 0%    8.32ms ± 1%   +106.15%  (p=0.000 n=8+9)

name           old alloc/op   new alloc/op   delta
Query/Types-8    5.69MB ± 0%   92.23MB ± 0%  +1520.17%  (p=0.000 n=10+8)
Query/Merge-8    74.9MB ± 0%    93.8MB ± 1%    +25.26%  (p=0.000 n=9+10)
Query/Range-8    16.3MB ± 0%    27.0MB ± 0%    +65.72%  (p=0.000 n=10+10)

name           old allocs/op  new allocs/op  delta
Query/Types-8     93.4k ± 0%     24.2k ± 0%    -74.13%  (p=0.000 n=10+9)
Query/Merge-8      301k ± 0%      204k ± 0%    -32.12%  (p=0.000 n=9+10)
Query/Range-8     57.7k ± 0%      8.4k ± 0%    -85.44%  (p=0.000 n=10+10)
```