@asubiotto 2022-10-20
Before merging, I'd like to add some logging/metrics.