@asubiotto 2022-10-24
I made some modifications (level 1 target size and size to split granule) and ran these new compaction changes on a local parca instance ingesting profiles from demo and it stabilized performance significantly. In the following graphs, I ran the old compactor until a block rotation and then ran the new compactor. The new compactor starts at 12:05 on the graph, so basically where all curves drop back down midway through:

Here is the p50 of Query (Merge/Single):
<img width="1407" alt="image" src="https://user-images.githubusercontent.com/10560359/197533902-623ec087-4d68-4ea5-b7c1-e0b24693813e.png">

p50 of QueryRange:
<img width="1396" alt="image" src="https://user-images.githubusercontent.com/10560359/197534016-e1e18504-f799-4833-90df-4f11cb4fab8e.png">

p90 latencies are also relatively more stable:
<img width="1402" alt="image" src="https://user-images.githubusercontent.com/10560359/197534105-3a5fc475-35f5-4914-b697-67c9f691cbd4.png">
<img width="1411" alt="image" src="https://user-images.githubusercontent.com/10560359/197534170-289e3b58-699a-46a2-b829-e93e37b5df1c.png">

The downside is that the active block is rotated earlier than with the old compactor since the old compactor recompacted everything.
<img width="1403" alt="image" src="https://user-images.githubusercontent.com/10560359/197534657-f78ba612-1694-4009-8cf7-4a6723176830.png">

In general, we also distribute the parts evenly to avoid excessive compactions:
<img width="1405" alt="image" src="https://user-images.githubusercontent.com/10560359/197534779-ff2f0c21-36fb-4fc0-9a8a-4694d8405b79.png">

This probably means that we could probably be doing more compactions while not affecting query latency/stability. I think I'm fine merging this as a first version and then iterating with the parameters to find what works for us.